+++ components/Modal/sub-components/Content.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const Content = styled.div`
+  font-size: 14px;
+  padding: 18px 24px;
+`;
+
+Content.displayName = 'Modal.Content';
+
+export default Content;